+++ packages/docs/.storybook/TableComponent.tsx
@@ -24,13 +24,13 @@ export const TableComponent = (props) => {
   const info = props.type.__docgenInfo;
   if (!info || !info.props) {
     return (
-      <ThemeProvider theme={Themes.sap_fiori3_light} contentDensity={ContentDensity.Compact}>
+      <ThemeProvider theme={Themes.sap_fiori_3} contentDensity={ContentDensity.Compact}>
         <Text>Unfortunately, there are no prop types available for this component.</Text>
       </ThemeProvider>
     );
   }
   return (
-    <ThemeProvider theme={Themes.sap_fiori3_light} contentDensity={ContentDensity.Compact}>
+    <ThemeProvider theme={Themes.sap_fiori_3} contentDensity={ContentDensity.Compact}>
       <Table
         columns={columns}
         rows={Object.values(info.props).map((componentInfo: any) => (
