+++ packages/fiori3/src/enums/Themes.ts
@@ -1,3 +1,3 @@
 export enum Themes {
-  sap_fiori3_light = 'sap_fiori3_light'
+  sap_fiori_3 = 'sap_fiori_3'
 }
