+++ packages/fiori3/src/interfaces/JSSTheme.ts
@@ -1,9 +1,9 @@
 import { ContentDensity } from '../enums/ContentDensity';
 import { Themes } from '../enums/Themes';
-import { sap_fiori3_light } from '@fiori-for-react/styles';
+import { sap_fiori_3 } from '@fiori-for-react/styles';
 
 export interface JSSTheme {
   theme: Themes;
   contentDensity: ContentDensity;
-  parameters: typeof sap_fiori3_light;
+  parameters: typeof sap_fiori_3;
 }
