+++ packages/styles/src/index.ts
@@ -3,7 +3,7 @@ import font72 from './72';
 import * as sap_belize from './sap_belize';
 import * as sap_belize_plus from './sap_belize_plus';
 import * as sap_belize_hcb from './sap_belize_hcb';
-import * as sap_fiori3_light from './sap_fiori3_light';
+import * as sap_fiori_3 from './sap_fiori_3';
 import * as spacing from './spacing';
 import { HSLColor } from './HSLColor';
 import { bootstrap } from './bootstrap';
@@ -17,7 +17,7 @@ export {
   sap_belize_hcb,
   spacing,
   HSLColor,
-  sap_fiori3_light,
+  sap_fiori_3,
   bootstrap,
   withStyles
 };
